done by sir

+++ ProjectBackend/src/main/java/org/isdb/ProjectBackend/repository/ReviewRepository.java
@@ -3,6 +3,6 @@ package org.isdb.ProjectBackend.repository;
 import org.isdb.ProjectBackend.model.Review;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface ReviewRepository extends JpaRepository<Review, Integer> {
-    // You can add custom query methods here if needed
+public interface ReviewRepository extends JpaRepository<Review, Long> {
+	// You can add custom query methods here if needed
 }
